fixed "error[E0308]: mismatched types    --> packages\blitz-winrt\src\event_conversion.rs:471:35     | 471 |         const VK_SHIFT_VAL: i32 = VK_SHIFT.0;     |                                   ^^^^^^^^^^ expected `i32`, found `u16`
error[E0308]: mismatched types
   --> packages\blitz-winrt\src\event_conversion.rs:472:37
    |
472 |         const VK_CONTROL_VAL: i32 = VK_CONTROL.0;
    |                                     ^^^^^^^^^^^^ expected `i32`, found `u16`

error[E0308]: mismatched types
   --> packages\blitz-winrt\src\event_conversion.rs:473:34
    |
473 |         const VK_MENU_VAL: i32 = VK_MENU.0;
    |                                  ^^^^^^^^^ expected `i32`, found `u16`

error[E0308]: mismatched types
   --> packages\blitz-winrt\src\surface_manager.rs:83:38
    |
83  |           let instance = Instance::new(wgpu::InstanceDescriptor {
    |  ________________________-------------_^
    | |                        |
    | |                        arguments to this function are incorrect
84  | |             backends: wgpu::Backends::DX12, // Use DX12 for SwapChainPanel compatibility
85  | |             flags: wgpu::InstanceFlags::default(),
86  | |             backend_options: wgpu::BackendOptions::default(),
87  | |         });
    | |_________^ expected `&InstanceDescriptor`, found `InstanceDescriptor`
    |"

+++ packages/blitz-winrt/src/event_conversion.rs
@@ -468,9 +468,9 @@ impl EventConverter {
     /// Updates modifier state based on key press/release.
     fn update_modifier_state_from_key(&mut self, virtual_key: u16, pressed: bool) {
         // Use const values to avoid snake_case warnings
-        const VK_SHIFT_VAL: i32 = VK_SHIFT.0;
-        const VK_CONTROL_VAL: i32 = VK_CONTROL.0;
-        const VK_MENU_VAL: i32 = VK_MENU.0;
+        const VK_SHIFT_VAL: i32 = VK_SHIFT.0 as i32;
+        const VK_CONTROL_VAL: i32 = VK_CONTROL.0 as i32;
+        const VK_MENU_VAL: i32 = VK_MENU.0 as i32;
         
         match VIRTUAL_KEY(virtual_key as i32) {
             key if key.0 == VK_SHIFT_VAL => self.modifier_state.shift = pressed,
